fix: disable multisig after remove signer with etd expires
In case the client side did not bundle remove signer with etd together with disable multisig with etd we need to disable multisig when the remove signer happens as it does not make sense to keep multisig in this scenario.

The user can still disable multisig with etd in this scenario but he will have to wait for ETD (4 days) - we would like to avoid that in case of client side bug.